// apps/planes_de_trabajo/src/app/core/models/profesor.model.ts
export interface Profesor {
  numIdentificacion: string;
  nombres: string;
  apellidos: string;
  tipoIdentificacion: string;
  facultad: string;
  programa: string;
  cargo: string;
  empresa: string;
  empleado: string;
  tipoDedicacion: string;
  secuencia: string;
  periodo: string;
  centroCosto: string;
  organizacion: string;
  estado: string;
  ordenPeriodo: string;
  cantidadFijo: number;
  fondo: string;
  fuenteFuncion: string;
  observaciones: string;
  categoria: string;
  nivelEducativo: string;
  escalafon: string;
  dedicacion: string;
  vinculacion: string;
  planDeTrabajo?: any;
}